services/extruct-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
extruct==0.18.0
python-dateutil==2.8.2
pydantic==2.5.3
